scripts/mail_handler/requirements.txt
google-api-python-client==2.86.0
google-auth==2.22.0
google-auth-oauthlib==1.2.0
pyyaml==6.0
python-dateutil==2.8.2
